feat: enabled savepoints for advanced-client-side-transaction-retries
To have ActiveRecord controll all the records and issue correct `before_committed!`, `comitted!` and `rolledback!` callbacks during serialisation errors and retries, we need a transaction instance that is able to execute all the savepoint statements on its own and at the right moment.
Otherwise we are unable to use the advanced-client-side-transaction-retries in a save way.

https://www.cockroachlabs.com/docs/stable/advanced-client-side-transaction-retries.html